updater/requirements.txt
xlrd
beautifulsoup4
requests
PyMySQL
python-dateutil